requirements.txt
fastapi
uvicorn
apscheduler
requests
python-dotenv
pydantic
